backend/requirements.txt
fastapi==0.127.0
uvicorn==0.40.0
python-multipart==0.0.21
faster-whisper==1.1.0
numpy==2.3.5
transformers
torch
sentencepiece
groq
python-dotenv
pymongo
bcrypt
python-jose[cryptography]
yt-dlp
pypdf
python-docx
email-validator
requests
